Setup integration_test file structure
Add a new integration test file within `tests/`. Also, edit the
.gitignore to ignore all files in the data/ directory except for a
.gitkeep files. The data/ directory will be used to house example html
files (or possible other data) to facilitate the integration testing.
